Move items models into Items models folder.
