+++ GuysNight.LethalCompanyMod.BalancedItems/Patches/StartOfRoundPatches.cs
@@ -1,5 +1,5 @@
 ﻿#pragma warning disable S1118
-using GuysNight.LethalCompanyMod.BalancedItems.Models;
+using GuysNight.LethalCompanyMod.BalancedItems.Models.Items;
 using GuysNight.LethalCompanyMod.BalancedItems.Models.Moons;
 using GuysNight.LethalCompanyMod.BalancedItems.Utilities;
 using HarmonyLib;
